fix for Sort function -- default arg -1 is a bit tricky with Variants and SafeEl! ak! also minor compile warning fix on imgproc

+++ src/temt/ta/ta_imgproc.cpp
@@ -1611,9 +1611,9 @@ bool taImageProc::RenderOccluderBorderColor_float(float_Matrix& img_data,
   }
   TwoDCoord img_size(img_data.dim(0), img_data.dim(1));
   TwoDCoord ll;
-  ll.x = img_size.x * llx;  ll.y = img_size.y * lly;
+  ll.x = (int)(img_size.x * llx);  ll.y = (int)img_size.y * lly;
   TwoDCoord ur;
-  ur.x = img_size.x * urx;  ur.y = img_size.y * ury;
+  ur.x = (int)(img_size.x * urx);  ur.y = (int)(img_size.y * ury);
   float oavg = img_data.FastEl(0,0); // assuming already has renderborder called
   for(int y = ll.y; y < ur.y; y++) {
     for(int x = ll.x; x < ur.x; x++) {
